# Ix: Improve the performance of Buffer exact/skip mode